Add new component and use call back
