+++ src/components/appComponents/Books.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import BookList from './booksComponents/BookList';
 import AddBookForm from './booksComponents/AddBooForm';
-import AddForm from './booksComponents/AddForm';
 
 function Books() {
   return (
     <div>
       <BookList />
       <AddBookForm />
-      <AddForm />
     </div>
   );
 }
