--- src/components/appComponents/booksComponents/AddForm.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import BookForm from './BookForm';
-
-function AddForm() {
-  return <BookForm />;
-}
-
-export default AddForm;
